Extended configurations to allow to add an icon for the applied configuration
