+++ components/bpmn-q/modeler-component/extensions/data-extension/rendering/DataFlowRenderer.js
@@ -1,8 +1,9 @@
 import BpmnRenderer from "bpmn-js/lib/draw/BpmnRenderer";
 import * as consts from '../Constants';
 import {attr as svgAttr} from 'tiny-svg';
-import {drawDataStoreSVG, drawTaskSVG} from "../../../common/util/RenderUtilities";
+import {drawDataElementSVG, drawTaskSVG} from "../../../common/util/RenderUtilities";
 import {getSVG} from "./DataFlowSVGMap";
+import {extractSVG} from '../../../editor/configurations/ConfigurationsUtil';
 
 export default class DataFlowRenderer extends BpmnRenderer {
 
@@ -14,14 +15,15 @@ export default class DataFlowRenderer extends BpmnRenderer {
       [consts.DATA_MAP_OBJECT]: function (self, parentGfx, element) {
         const task = self.renderer('bpmn:DataObject')(parentGfx, element);
 
-        drawDataStoreSVG(parentGfx, getSVG(consts.DATA_TYPE_DATA_MAP_OBJECT));
+        let svg = extractSVG(element.businessObject.get('configsIcon')) || getSVG(consts.DATA_TYPE_DATA_MAP_OBJECT);
+        drawDataElementSVG(parentGfx, svg);
 
         return task;
       },
       [consts.DATA_STORE_MAP]: function (self, parentGfx, element) {
         const task = self.renderer('bpmn:DataStoreReference')(parentGfx, element);
 
-        drawDataStoreSVG(parentGfx, getSVG(consts.DATA_TYPE_DATA_STORE_MAP));
+        drawDataElementSVG(parentGfx, getSVG(consts.DATA_TYPE_DATA_STORE_MAP));
 
         return task;
       },
